Move Commit to lib
