+++ src/main.rs
@@ -116,7 +116,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     let mut commits = commits.as_str();
     let mut parsed_commits = vec![];
     while !commits.is_empty() {
-        let result = parse_commit(commits);
+        let result = commit_analyzer::commit::Commit::new(commits);
         match result {
             Ok((commit, remainder)) => {
                 commits = remainder;
@@ -128,7 +128,7 @@ fn main() -> Result<(), Box<dyn Error>> {
             }
         }
     }
-    let mut last_time = None;
+    let mut last_time: std::option::Option<chrono::DateTime<chrono::FixedOffset>> = None;
     let mut duration = chrono::Duration::zero();
     let filter = commit_analyzer::filter::Filter::new(&matches);
     let mut commit_count = 0;
@@ -138,20 +138,20 @@ fn main() -> Result<(), Box<dyn Error>> {
         if filter.matches(&commit) {
             commit_count += 1;
             commits_per_day
-                .entry(commit.date.date())
+                .entry(commit.date().date())
                 .or_insert(0)
                 .add_assign(1);
             loc_per_day
-                .entry(commit.date.date())
+                .entry(commit.date().date())
                 .or_insert(0)
                 .add_assign(commit.loc(&filter));
             if let Some(last_time) = last_time {
-                let diff: chrono::Duration = commit.date - last_time;
+                let diff: chrono::Duration = *commit.date() - last_time;
                 if diff.num_hours() <= max_diff_hours as i64 {
                     duration = duration + diff;
                 }
             }
-            last_time = Some(commit.date);
+            last_time = Some(*commit.date());
             if is_verbose {
                 println!("{:#?}", commit);
             }
@@ -177,132 +177,4 @@ fn main() -> Result<(), Box<dyn Error>> {
     }
 
     Ok(())
-}
-
-#[derive(Debug)]
-enum CommitParseError {
-    CommitMissing,
-    AuthorMissing,
-    DateMissing,
-    AuthorFailed(commit_analyzer::author::AuthorParseError),
-    DateFailed(chrono::ParseError),
-    LocSyntaxError,
-    LocFailed(commit_analyzer::loc::LocParseError),
-    Unknown,
-}
-
-fn parse_commit(commit: &str) -> Result<(Commit, &str), CommitParseError> {
-    let (commit, remainder) = commit
-        .strip_prefix("commit")
-        .ok_or(CommitParseError::CommitMissing)?
-        .split_once('\n')
-        .ok_or(CommitParseError::CommitMissing)?;
-    let commit = commit.trim();
-    let (merge, remainder) = if let Some(it) = remainder
-        .strip_prefix("Merge: ")
-        .map(|s| s.split_once('\n').ok_or(CommitParseError::Unknown))
-    {
-        let (merge, remainder) = it?;
-        (Some(merge.trim().to_owned()), remainder)
-    } else {
-        (None, remainder)
-    };
-    let (author, remainder) = remainder
-        .strip_prefix("Author: ")
-        .ok_or(CommitParseError::AuthorMissing)?
-        .split_once('\n')
-        .ok_or(CommitParseError::AuthorMissing)?;
-    let (date, mut remainder) = remainder
-        .strip_prefix("Date:   ")
-        .ok_or(CommitParseError::DateMissing)?
-        .split_once('\n')
-        .ok_or(CommitParseError::DateMissing)?;
-
-    let mut message = String::new();
-    let mut remainder_result = remainder;
-    if let Some(it) = remainder.strip_prefix('\n') {
-        remainder = it;
-    }
-    let mut space_count = 0;
-    let mut increase_space_count = true;
-    for (index, char) in remainder.char_indices() {
-        if increase_space_count {
-            if char == ' ' {
-                space_count += 1;
-            } else {
-                increase_space_count = false;
-            }
-        } else if space_count == 4 && char == '\n' {
-            increase_space_count = true;
-            space_count = 0;
-        } else if space_count < 4 {
-            break;
-        }
-        // This removes the char from remainder, before adding it to the
-        // message.
-        remainder_result = &remainder[index + char.len_utf8()..];
-        if !increase_space_count {
-            message.push(char);
-        }
-    }
-    let message = message.trim();
-
-    let mut locs = vec![];
-    loop {
-        if remainder_result.is_empty() {
-            break;
-        }
-        let (loc, remainder) = remainder_result
-            .split_once('\n')
-            .ok_or(CommitParseError::LocSyntaxError)?;
-        if loc.is_empty() {
-            // We still need to consume the last line feed, otherwise the parser
-            // will fail on the last commit.
-            remainder_result = remainder;
-            break;
-        } else if loc.starts_with("commit") {
-            break;
-        }
-        locs.push(commit_analyzer::loc::LocDiff::parse(loc).map_err(CommitParseError::LocFailed)?);
-        remainder_result = remainder;
-        if let Some(remainder) = remainder_result.strip_prefix('\n') {
-            remainder_result = remainder;
-            break;
-        }
-    }
-
-    let commit = Commit {
-        commit: commit.into(),
-        merge,
-        author: commit_analyzer::author::Author::new(author)
-            .map_err(CommitParseError::AuthorFailed)?,
-        date: chrono::DateTime::parse_from_str(date, "%a %b %e %T %Y %z")
-            .map_err(CommitParseError::DateFailed)?,
-        message: message.into(),
-        locs,
-    };
-    Ok((commit, remainder_result))
-}
-
-#[derive(Debug)]
-pub struct Commit {
-    commit: String,
-    /// This field is currently unused, if this changes, you can delete this
-    /// comment and the allow(dead_code) tag.
-    #[allow(dead_code)]
-    merge: Option<String>,
-    author: commit_analyzer::author::Author,
-    date: chrono::DateTime<chrono::FixedOffset>,
-    message: String,
-    locs: Vec<commit_analyzer::loc::LocDiff>,
-}
-
-impl Commit {
-    pub fn loc(&self, filter: &commit_analyzer::filter::Filter) -> i64 {
-        self.locs
-            .iter()
-            .filter(|l| filter.check_loc(l))
-            .map(|l| l.loc())
-            .sum()
-    }
 }
